Amiba

+++ database/migrations/2020_03_23_003302_favorites.php
@@ -14,9 +14,9 @@ class Favorites extends Migration
     public function up()
     {
         Schema::create('favorites', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id')->unsigned();
-            $table->integer('question_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id')->unsigned();
+            $table->unsignedBigInteger('question_id')->unsigned();
             $table->timestamps();
             $table->foreign('user_id')
                   ->references('id')
